+++ src/main/java/com/atguigu/cloud/iotcloudspring/mapper/ICMapper.java
@@ -15,22 +15,22 @@ public interface ICMapper {
     List<ProjectAdd> GetProject(String username);
 
     //添加projectadd的数据
-    int AddProject(ProjectAdd projectAdd);
+    Long AddProject(ProjectAdd projectAdd);
 
     //查询用户名是否存在
     users selectUserByUsername(String username);
 
     //注册用户
-    int RegisteredUser(users user);
+    Long RegisteredUser(users user);
 
     //修改密码
-    int updatePassword(@Param("username") String username, @Param("newPassword") String newPassword);
+    Long updatePassword(@Param("username") String username, @Param("newPassword") String newPassword);
 
     // 插入用户-项目关联记录
-    int insertUserProject(UserProject userProject);
+    Long insertUserProject(UserProject userProject);
 
     // 根据用户ID查询该用户加入的所有项目关联记录
-    List<UserProject> selectByUserId(@Param("userid") Integer userid);
+    List<UserProject> selectByUserId(@Param("userid") Long userid);
 
     // 根据项目创建者（admin 字段）查询该用户创建的所有项目
     List<ProjectAdd> selectProjectsByAdmin(@Param("projectadministrator") String projectadministrator);
@@ -39,15 +39,15 @@ public interface ICMapper {
     ProjectAdd selectProjectByName(@Param("projectname") String projectname);
 
     // 删除用户-项目关联记录（退出项目）
-    int deleteUserProject(@Param("userid") Integer userid, @Param("projectid") Integer projectid);
+    Long deleteUserProject(@Param("userid") Long userid, @Param("projectid") Long projectid);
 
     // 根据项目ID查询项目
-    List<ProjectAdd> selectJoinedProjectsByUser(@Param("userid") Integer userid);
+    List<ProjectAdd> selectJoinedProjectsByUser(@Param("userid") Long userid);
 
     // 根据用户ID和项目ID查询用户-项目关联记录
-    UserProject selectUserProjectByUserIdAndProjectId(@Param("userid") Integer userid,
-                                                      @Param("projectid") Integer projectid);
+    UserProject selectUserProjectByUserIdAndProjectId(@Param("userid") Long userid,
+                                                      @Param("projectid") Long projectid);
 
     // 根据项目ID查询项目
-    ProjectAdd selectProjectById(@Param("id") Integer id);
+    ProjectAdd selectProjectById(@Param("id") Long id);
 }
