+++ src/main/java/com/atguigu/cloud/iotcloudspring/pojo/device/Device.java
@@ -10,9 +10,9 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Device {
-    private Integer id;
-    private Integer projectid;
-    private Integer devicetypeid;
+    private Long id;
+    private Long projectid;
+    private Long devicetypeid;
     private String devicename;
     private String devicekey;
     private String devicecommunication;
