+++ src/main/java/com/atguigu/cloud/iotcloudspring/service/DeviceStatusService.java
@@ -4,5 +4,5 @@ import com.atguigu.cloud.iotcloudspring.DTO.Device.Response.DeviceStatusResponse
 import com.atguigu.cloud.iotcloudspring.pojo.device.Device;
 
 public interface DeviceStatusService {
-    DeviceStatusResponse updateOrInsertDeviceStatus(Integer id, String devicestatus);
+    DeviceStatusResponse updateOrInsertDeviceStatus(Long id, String devicestatus);
 }
